[WIP] Fix haproxy_server_max_sessions test
This assertion fails when the router reloads. It may be better to mark
this test as serial. This is an experiment to retry the assertion. The
expectation is that haxproxy_server_max_sessions needs to be greater
than zero over two successive scrapes once the first scrape shows an
increase. However, this assertion can still fail depending on when the
metrics are refreshed and when the router reloads, which is every 5s.