test/extended/router/idle: address flakes/failures seen in CI
Fixes for failures/flakes seen in CI runs:

- increase timeout for polling operations to 5 mins
- increase timeout for URL get operations to 2 mins

- wait for PODs to be running after the fixture has been created
- wait for PODs to be running after service has been unidled

This also switches from using URLTester to making a direct connection
to the route's hostname. This reduces the need for an execpod and will
make the test run faster. It is also a better end-2-end test.

It additionally dumps all the resources from the test (and only from
the test namespace) to aid debugging if it continues to flake in CI.